refactor: оптимизация Reviews
