+++ src/pages/offer/components/reviews/reviews.tsx
@@ -1,10 +1,10 @@
 import { memo } from 'react';
-import { useAppSelector } from '../../../hooks';
-import { authorizationSelectors } from '../../../store/slices/authorization';
-import { offerSelectors } from '../../../store/slices/offer';
-import { getPluralForm } from '../../../util';
-import Review from './review';
+import { useAppSelector } from '../../../../hooks';
+import { authorizationSelectors } from '../../../../store/slices/authorization';
+import { offerSelectors } from '../../../../store/slices/offer';
+import { getPluralForm } from '../../../../util';
 import ReviewsForm from './reviews-form';
+import ReviewsList from './reviews-list';
 
 type ReviewsProps = {
   idDetailedOffer: string;
@@ -22,11 +22,7 @@ function Reviews({ idDetailedOffer }: ReviewsProps): JSX.Element {
         <h2 className="reviews__title">
           {getPluralForm('Review', reviewsOffer.length)} · <span className="reviews__amount">{reviewsOffer.length}</span>
         </h2>}
-      <ul className="reviews__list">
-        {reviewsOffer.map((reviewOffer) => (
-          <Review reviewOffer={reviewOffer} key={reviewOffer.id} />
-        ))}
-      </ul>
+      <ReviewsList reviewsOffer={reviewsOffer} />
       {isAuth && <ReviewsForm offerId={idDetailedOffer} />}
     </section>
   );
